[CL-3715] Budgeting e2e test
